+++ sdk/schemaregistry/azure-schemaregistry/azure/schemaregistry/_schema_registry_client.py
@@ -91,8 +91,8 @@ class SchemaRegistryClient(object):
     def register_schema(
         self,
         group_name,
-        schema_name,
-        schema_definition,
+        name,
+        definition,
         format,
         **kwargs  # pylint:disable=redefined-builtin
     ):
@@ -103,8 +103,8 @@ class SchemaRegistryClient(object):
         schema is created at latest version + 1.
 
         :param str group_name: Schema group under which schema should be registered.
-        :param str schema_name: Name of schema being registered.
-        :param str schema_definition: String representation of the schema being registered.
+        :param str name: Name of schema being registered.
+        :param str definition: String representation of the schema being registered.
         :param format: Format for the schema being registered.
          For now Avro is the only supported schema format by the service.
         :type format: Union[str, SchemaFormat]
@@ -130,8 +130,8 @@ class SchemaRegistryClient(object):
         http_request_kwargs = get_http_request_kwargs(kwargs)
         request = schema_rest.build_register_request(
             group_name=group_name,
-            schema_name=schema_name,
-            content=schema_definition,
+            schema_name=name,
+            content=definition,
             content_type=kwargs.pop(
                 "content_type", "application/json; serialization={}".format(format)
             ),
@@ -173,8 +173,8 @@ class SchemaRegistryClient(object):
     def get_schema_properties(
         self,
         group_name,
-        schema_name,
-        schema_definition,
+        name,
+        definition,
         format,
         **kwargs  # pylint:disable=redefined-builtin
     ):
@@ -184,8 +184,8 @@ class SchemaRegistryClient(object):
         as matched by schema definition comparison.
 
         :param str group_name: Schema group under which schema should be registered.
-        :param str schema_name: Name of schema being registered.
-        :param str schema_definition: String representation of the schema being registered.
+        :param str name: Name of schema being registered.
+        :param str definition: String representation of the schema being registered.
         :param format: Format for the schema being registered.
         :type format: Union[str, SchemaFormat]
         :rtype: ~azure.schemaregistry.SchemaProperties
@@ -210,8 +210,8 @@ class SchemaRegistryClient(object):
         http_request_kwargs = get_http_request_kwargs(kwargs)
         request = schema_rest.build_query_id_by_content_request(
             group_name=group_name,
-            schema_name=schema_name,
-            content=schema_definition,
+            schema_name=name,
+            content=definition,
             content_type=kwargs.pop(
                 "content_type", "application/json; serialization={}".format(format)
             ),